requirements.txt
streamlit==1.22.0
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
pillow==9.5.0
requests==2.28.2
torch==1.13.1
torchvision==0.14.1